コマンドプロンプト表示処理のファイル分割
